examples/with_db/requirements.txt
python-fasthtml
fastlite
